Reorganize settings and merge AI configuration
Groups settings by workflow, merges AI usage into the AI tab, and adds settings modal close affordances.
Includes passing formatting, lint, build, and test checks.
