+++ Muxy/Views/Settings/AIAssistantSettingsView.swift
@@ -74,6 +74,10 @@ struct AIAssistantSettingsView: View {
                     onReset: { prPrompt = "" }
                 )
             }
+
+            SettingsSection("AI Usage", showsDivider: false) {
+                AIUsageSettingsView()
+            }
         }
     }
