+++ Muxy/Views/Settings/GeneralSettingsView.swift
@@ -1,26 +1,10 @@
-import AppKit
 import SwiftUI
 
 struct GeneralSettingsView: View {
-    @AppStorage(GeneralSettingsKeys.autoExpandWorktreesOnProjectSwitch)
-    private var autoExpandWorktrees = false
-    @AppStorage(GeneralSettingsKeys.defaultWorktreeParentPath)
-    private var defaultWorktreeParentPath = ""
-    @AppStorage(GeneralSettingsKeys.fileTreeSource)
-    private var fileTreeSourceRaw = FileTreeSourcePreference.defaultValue.rawValue
-    @AppStorage(TabCloseConfirmationPreferences.confirmRunningProcessKey)
-    private var confirmRunningProcess = true
-    @AppStorage(ProjectLifecyclePreferences.keepOpenWhenNoTabsKey)
-    private var keepProjectsOpenWhenNoTabs = false
-    @AppStorage(ProjectPickerPreferences.storageKey)
-    private var projectPickerModeRaw = ProjectPickerMode.custom.rawValue
     @AppStorage(UpdateChannel.storageKey)
     private var updateChannelRaw = UpdateChannel.stable.rawValue
     @AppStorage(QuitConfirmationPreferences.confirmQuitKey)
     private var confirmQuit = true
-    @AppStorage(GeneralSettingsKeys.autoCopyTerminalSelection)
-    private var autoCopyTerminalSelection = false
-    @State private var projectPickerDefaultLocationSettings = ProjectPickerDefaultLocationSettingsModel()
     @State private var sentry = SentryService.shared
 
     var body: some View {
@@ -39,84 +23,6 @@ struct GeneralSettingsView: View {
                     .labelsHidden()
                     .frame(width: SettingsMetrics.controlWidth, alignment: .trailing)
                 }
-            }
-
-            SettingsSection(
-                "Sidebar",
-                footer: "Automatically reveal worktrees when you switch to a project."
-            ) {
-                SettingsToggleRow(
-                    label: "Auto-expand worktrees on project switch",
-                    isOn: $autoExpandWorktrees
-                )
-            }
-
-            SettingsSection(
-                "File Tree",
-                footer: "When set to the active terminal, the file tree follows the working directory of "
-                    + "the active terminal tab. If there is no active terminal, it keeps the last known path."
-            ) {
-                SettingsRow("Root directory") {
-                    Picker("", selection: $fileTreeSourceRaw) {
-                        ForEach(FileTreeSourcePreference.allCases) { source in
-                            Text(source.title).tag(source.rawValue)
-                        }
-                    }
-                    .labelsHidden()
-                    .frame(width: SettingsMetrics.controlWidth, alignment: .trailing)
-                }
-            }
-
-            SettingsSection(
-                "Projects",
-                footer: projectsFooter
-            ) {
-                SettingsRow("Muxy Picker") {
-                    Picker("", selection: $projectPickerModeRaw) {
-                        ForEach(ProjectPickerMode.allCases) { mode in
-                            Text(mode.label).tag(mode.rawValue)
-                        }
-                    }
-                    .labelsHidden()
-                    .frame(width: SettingsMetrics.controlWidth, alignment: .trailing)
-                }
-
-                if projectPickerMode == .custom {
-                    ProjectPickerDefaultLocationSettingsView(
-                        model: projectPickerDefaultLocationSettings,
-                        pickerModeRaw: projectPickerModeRaw
-                    )
-                }
-
-                SettingsToggleRow(
-                    label: "Keep projects open after closing the last tab",
-                    isOn: $keepProjectsOpenWhenNoTabs
-                )
-            }
-
-            SettingsSection(
-                "Worktrees",
-                footer: "Muxy creates a project-named subfolder inside this folder. "
-                    + "Projects can still override this from the new worktree dialog."
-            ) {
-                worktreeLocationControl
-            }
-
-            SettingsSection(
-                "Terminal",
-                footer: "When enabled, releasing the mouse after selecting text in the terminal copies it to the clipboard."
-            ) {
-                SettingsToggleRow(
-                    label: "Auto-copy selected text",
-                    isOn: $autoCopyTerminalSelection
-                )
-            }
-
-            SettingsSection("Tabs") {
-                SettingsToggleRow(
-                    label: "Confirm before closing a tab with a running process",
-                    isOn: $confirmRunningProcess
-                )
             }
 
             SettingsSection("Quit", showsDivider: sentry.hasDSN) {
@@ -157,84 +63,5 @@ struct GeneralSettingsView: View {
                 UpdateService.shared.channel = newValue
             }
         )
-    }
-
-    private var projectPickerMode: ProjectPickerMode {
-        ProjectPickerMode(rawValue: projectPickerModeRaw) ?? .custom
-    }
-
-    private var projectsFooter: String {
-        if projectPickerMode == .custom {
-            return "Muxy Picker starts in this default location. Use App Default to reset it. "
-                + "Projects can stay in the sidebar after closing their last tab."
-        }
-        return "Muxy Picker can use Finder or Muxy's picker. Projects can stay in the sidebar after closing their last tab."
-    }
-
-    private var defaultWorktreeLocationText: String {
-        defaultWorktreeParentPath.isEmpty ? "Muxy App Support" : defaultWorktreeParentPath
-    }
-
-    private var worktreeLocationControl: some View {
-        VStack(alignment: .leading, spacing: 7) {
-            Text("Default path for new worktrees")
-                .font(.system(size: SettingsMetrics.labelFontSize))
-
-            HStack(alignment: .center, spacing: 8) {
-                pathDisplay
-                    .layoutPriority(1)
-
-                Button("Choose Folder...") {
-                    chooseDefaultWorktreeParentPath()
-                }
-                .fixedSize(horizontal: true, vertical: false)
-
-                Button("Use App Default") {
-                    defaultWorktreeParentPath = ""
-                }
-                .fixedSize(horizontal: true, vertical: false)
-                .disabled(defaultWorktreeParentPath.isEmpty)
-            }
-        }
-        .padding(.horizontal, SettingsMetrics.horizontalPadding)
-        .padding(.vertical, SettingsMetrics.rowVerticalPadding)
-    }
-
-    private var pathDisplay: some View {
-        HStack(spacing: 7) {
-            Image(systemName: defaultWorktreeParentPath.isEmpty ? "internaldrive" : "folder")
-                .font(.system(size: 11, weight: .medium))
-                .foregroundStyle(SettingsStyle.mutedForeground)
-                .frame(width: 15)
-
-            Text(defaultWorktreeLocationText)
-                .font(.system(size: SettingsMetrics.footnoteFontSize, design: .monospaced))
-                .foregroundStyle(defaultWorktreeParentPath.isEmpty ? SettingsStyle.mutedForeground : SettingsStyle.foreground)
-                .lineLimit(1)
-                .truncationMode(.middle)
-                .textSelection(.enabled)
-                .frame(maxWidth: .infinity, alignment: .leading)
-        }
-        .padding(.horizontal, 9)
-        .frame(minWidth: 170, maxWidth: .infinity, alignment: .leading)
-        .frame(height: 22)
-        .background(SettingsStyle.surface, in: RoundedRectangle(cornerRadius: 6))
-        .overlay(
-            RoundedRectangle(cornerRadius: 6)
-                .stroke(SettingsStyle.border, lineWidth: 1)
-        )
-    }
-
-    private func chooseDefaultWorktreeParentPath() {
-        let panel = NSOpenPanel()
-        panel.canChooseDirectories = true
-        panel.canChooseFiles = false
-        panel.allowsMultipleSelection = false
-        panel.message = "Select the default folder for new worktrees"
-        if let path = WorktreeLocationResolver.normalizedPath(defaultWorktreeParentPath) {
-            panel.directoryURL = URL(fileURLWithPath: path, isDirectory: true)
-        }
-        guard panel.runModal() == .OK, let url = panel.url else { return }
-        defaultWorktreeParentPath = url.path
     }
 }
